:recycle: refactor: Refatorada implementação de Cliente

+++ src/main/java/br/dev/paulowolfgang/gestao_apolices/service/impl/ClienteServiceImpl.java
@@ -1,14 +1,15 @@
 package br.dev.paulowolfgang.gestao_apolices.service.impl;
 
+import br.dev.paulowolfgang.gestao_apolices.dto.mapper.ClienteMapper;
+import br.dev.paulowolfgang.gestao_apolices.dto.request.ClienteRequestDto;
+import br.dev.paulowolfgang.gestao_apolices.dto.response.ClienteResponseDto;
 import br.dev.paulowolfgang.gestao_apolices.entity.Cliente;
-import br.dev.paulowolfgang.gestao_apolices.entity.ClienteFisico;
-import br.dev.paulowolfgang.gestao_apolices.entity.ClienteJuridico;
+import br.dev.paulowolfgang.gestao_apolices.entity.Usuario;
 import br.dev.paulowolfgang.gestao_apolices.repository.IClienteRepository;
 import br.dev.paulowolfgang.gestao_apolices.service.IClienteService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ClienteServiceImpl implements IClienteService {
@@ -20,54 +21,49 @@ public class ClienteServiceImpl implements IClienteService {
     }
 
     @Override
-    public Cliente salvar(Cliente cliente) {
-        return clienteRepository.save(cliente);
+    public ClienteResponseDto salvar(ClienteRequestDto request) {
+        Cliente cliente = ClienteMapper.converter(request);
+
+        // Simulação de um usuário temporário com ID fixo
+        Long idUsuarioSimulado = 1L;
+        Usuario usuarioSimulado = new Usuario();
+        usuarioSimulado.setId(idUsuarioSimulado);
+        cliente.setUsuario(usuarioSimulado);
+
+        cliente = clienteRepository.save(cliente);
+        return ClienteMapper.converter(cliente);
     }
 
     @Override
-    public Optional<Cliente> buscarPorId(Long id) {
-        return clienteRepository.findById(id);
+    public ClienteResponseDto buscarPorId(Long id) {
+        Cliente cliente = clienteRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Cliente não encontrado para o ID: " + id));
+        return ClienteMapper.converter(cliente);
     }
 
     @Override
-    public List<Cliente> listarTodos() {
-        return clienteRepository.findAll();
+    public List<ClienteResponseDto> listarTodos() {
+        List<Cliente> clientes = clienteRepository.findAll();
+        return clientes.stream()
+                .map(ClienteMapper::converter)
+                .toList();
     }
 
     @Override
-    public Cliente atualizar(Long id, Cliente clienteAtualizado) {
-        return clienteRepository.findById(id)
-                .map(clienteExistente -> {
+    public ClienteResponseDto atualizar(Long id, ClienteRequestDto clienteRequestDto) {
+        Cliente cliente = clienteRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Cliente não encontrado para o ID: " + id));
+        ClienteMapper.copiarParaPropriedades(clienteRequestDto, cliente);
+        cliente = clienteRepository.save(cliente);
 
-                    clienteExistente.setEmail(clienteAtualizado.getEmail());
-                    clienteExistente.setEndereco(clienteAtualizado.getEndereco());
-                    clienteExistente.setTelefone(clienteAtualizado.getTelefone());
-
-                    if (clienteExistente instanceof ClienteFisico && clienteAtualizado instanceof ClienteFisico) {
-                        ClienteFisico clienteFisicoExistente = (ClienteFisico) clienteExistente;
-                        ClienteFisico clienteFisicoAtualizado = (ClienteFisico) clienteAtualizado;
-
-                        clienteFisicoExistente.setNome(clienteFisicoAtualizado.getNome());
-                        clienteFisicoExistente.setCpf(clienteFisicoAtualizado.getCpf());
-                        clienteFisicoExistente.setDataNascimento(clienteFisicoAtualizado.getDataNascimento());
-                    } else if (clienteExistente instanceof ClienteJuridico && clienteAtualizado instanceof ClienteJuridico) {
-                        ClienteJuridico clienteJuridicoExistente = (ClienteJuridico) clienteExistente;
-                        ClienteJuridico clienteJuridicoAtualizado = (ClienteJuridico) clienteAtualizado;
-
-                        clienteJuridicoExistente.setNomeFantasia(clienteJuridicoAtualizado.getNomeFantasia());
-                        clienteJuridicoExistente.setRazaoSocial(clienteJuridicoAtualizado.getRazaoSocial());
-                        clienteJuridicoExistente.setCnpj(clienteJuridicoAtualizado.getCnpj());
-                        clienteJuridicoExistente.setDataAbertura(clienteJuridicoAtualizado.getDataAbertura());
-                    } else {
-                        throw new IllegalArgumentException("Tipo de cliente inválido ou incompatível para atualização.");
-                    }
-
-                    return clienteRepository.save(clienteExistente);
-                }).orElseThrow(() -> new RuntimeException("Cliente não encontrado"));
+        return ClienteMapper.converter(cliente);
     }
 
     @Override
     public void remover(Long id) {
+        if (!clienteRepository.existsById(id)) {
+            throw new RuntimeException("Cliente não encontrado para o ID: " + id);
+        }
         clienteRepository.deleteById(id);
     }
 }
